+++ src/s_sound.h
@@ -104,7 +104,7 @@ void S_StartMusic(int music_id);
 
 // Start music using <music_id> from sounds.h,
 //  and set whether looping
-void S_ChangeMusic(int music_id, dboolean looping, dboolean cheating, dboolean mapstart);
+void S_ChangeMusic(int music_id, dboolean looping, dboolean allowrestart, dboolean mapstart);
 
 // Stops the music fer sure.
 void S_StopMusic(void);