+++ releasenotes.md
@@ -35,6 +35,7 @@
 * The color of player messages in the alternate widescreen HUD is now correct when the `r_textures` CVAR is `off` and a `COLORMAP` lump that replaces colormap 32 is present in the current PWAD.
 * A bug has been fixed whereby the menu could appear with a corrupted palette during startup in some rare instances.
 * The red palette effect when the player is injured will now be allowed to fade out when the console is opened.
+* The `play` CCMD can now be used to restart music.
 
 ---